// src/types/enum.ts
export enum ExtraStatus {
  PendingApproval = 'Pending approval',
  Approved = 'Approved',
  Unapproved = 'Unapproved',
  NotInUse = 'Not in use',
  CoolingDown = 'Cooling down',
  Pending = 'Pending',
  Syncing = 'Syncing',
  Available = 'Available',
  Unavailable = 'Unavailable',
  Active = 'Active',
  Inactive = 'Inactive',
  Archived = 'Archived',
  Attention = 'Attention',
  Rejected = 'Rejected',
  Failed = 'Failed',
  Uploading = 'Uploading',
  Assigning = 'assigning',
  Finished = 'finished',
  Initial = 'initial',
}